+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Atmo France - Use new geocoding services to fix refresh issue for some locations
 - China - Fix nowcasting refresh silently failing
 - FMI - Added as a new forecast, current, air quality, alert, normals source for Finland (@chunshek). Users on devices with hardened memory allocator (`hardened_malloc`) enabled, such as GrapheneOS may want to disable the feature for the app if they find the refresh to be too slow.
+- Gadgetbridge - Improved weather codes mapping
 - Infoplaza - Added as a new worldwide forecast, current, minutely and normals source (@SKBotNL)
 - KNMI - Add textual representation of weather conditions (@willem640)
 - Météo-France/OpenWeather - Improve total precipitation computing (@chunshek)